fixed errors added method update

+++ src/main/java/com/repositories/DecoratorRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DecoratorRepository extends JpaRepository<BlackWhiteListEntity,Integer> {
+public interface DecoratorRepository extends JpaRepository<BlackWhiteListEntity,Long> {
 }
