+++ src/main/java/com/repositories/LockedNumberRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LockedNumberRepository extends JpaRepository<LockedNumberEntity,Integer> {
+public interface LockedNumberRepository extends JpaRepository<LockedNumberEntity,Long> {
 }
