+++ src/main/java/com/repositories/SubscriberRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SubscriberRepository extends JpaRepository<SubscriberEntity,Integer> {
+public interface SubscriberRepository extends JpaRepository<SubscriberEntity,Long> {
 }
